+++ app/forms/transfer/task/handover_task_form.rb
@@ -1,0 +1,5 @@
+class Transfer::Task::HandoverTaskForm < BaseOptionalTaskForm
+  attribute :review, :boolean
+  attribute :notes, :boolean
+  attribute :meeting, :boolean
+end